frame by frame mode, Wave maneger nästan klart

+++ Space_invaders_01/PhalanxPreset.cs
@@ -11,11 +11,11 @@ namespace Space_invaders_01
         public Enemy_type[] rows {  get; private set; }
         public int nr_of_collums { get; private set; }
         public int centerX { get; private set; }
-        public int row_space { get; private set; }
-        public int collum_space { get; private set; }
+        public float row_space { get; private set; }
+        public float collum_space { get; private set; }
         public int ad_speed { get; private set; }
 
-        public PhalanxPreset(Enemy_type[] _rows, int _nr_of_collums, int _centerX, int _row_space, int _collum_space, int _ad_speed)
+        public PhalanxPreset(Enemy_type[] _rows, int _nr_of_collums, int _centerX, float _row_space, float _collum_space, int _ad_speed)
         {
             this.rows = _rows;
             this.nr_of_collums = _nr_of_collums;
